🐛 FIX: Logout issue try 2
