+++ config/config-password.js
@@ -10,14 +10,11 @@ const params = {
 
 passport.use(
     new JwtStrategy(params, async function (payload, done) {
-        console.log("Payload JWT:", payload);
         try {
             const user = await User.findById(payload.id);
             if (!user) {
-                console.log("User not found");
                 return done(null, false);
             }
-            console.log("User found:", user);
             return done(null, user);
         } catch (err) {
             console.error("Error in JWT strategy:", err);
