+++ routes/api/users.js
@@ -2,7 +2,7 @@ const express = require("express");
 const User = require("../../models/schemaUser");
 require("dotenv").config();
 const jwt = require("jsonwebtoken");
-const auth = require("./contacts.js");
+const auth = require("../../auth/auth");
 
 const secret = process.env.AUTH_SECRET;
 
@@ -23,13 +23,13 @@ router.post("/signup", async (req, res, next) => {
         await newUser.save();
         return res
             .status(201)
-            .json({ message: `Registration succesful for ${email}` });
+            .json({ message: `Registration successful: user ${email}` });
     } catch (error) {
         next(error);
     }
 });
 
-router.post("/login", auth, async (req, res, next) => {
+router.post("/login", async (req, res, next) => {
     try {
         const { email, password } = req.body;
         const user = await User.findOne({ email });
@@ -51,7 +51,7 @@ router.post("/login", auth, async (req, res, next) => {
             user.token = token;
             await user.save();
             return res.status(200).json({
-                message: `Login succesful for ${email}`,
+                message: `Login succesful: user ${email}`,
                 data: { token },
             });
         }
@@ -63,12 +63,11 @@ router.post("/login", auth, async (req, res, next) => {
 router.get("/logout", auth, async (req, res, next) => {
     try {
         const user = req.user;
-        user.token = null;
-        await user.save();
-
         if (!user) {
             return res.status(401).json({ message: "User not authorized!" });
         }
+        user.token = null;
+        await user.save();
         return res.status(200).json({ message: "Logout successful" });
     } catch (error) {
         next(error);
@@ -79,10 +78,6 @@ router.get("/current", auth, async (req, res, next) => {
     try {
         const user = req.user;
 
-        console.log("Current user:", user);
-        console.log("Received contact ID:", req.params.contactId);
-        console.log("Headers:", req.headers);
-        console.log("User in current route:", req.user);
         if (!user) {
             return res.status(401).json({ message: "User not authorized!" });
         }
